requirements.txt
accelerate==0.18.0
apex==0.9.10dev
azureml==0.2.7
beautifulsoup4==4.11.2
classy_vision==0.7.0
clearml==1.12.1
codecarbon==2.2.7
comet_ml==3.33.7
cookiecutter==2.2.3
dall_e==0.1
Django==4.2.1
emoji==2.7.0
fairscale==0.4.13
fairseq==0.12.2
faiss==1.5.3
fastapi==0.100.1
filelock==3.8.2
flax==0.7.1
ftfy==6.1.1
fugashi==1.2.1
gdown==4.7.1
GitPython==3.1.31
GitPython==3.1.32
gluonnlp==0.10.0
h5py==3.8.0
haiku==0.01
huggingface_hub==0.16.4
huggingface_hub==0.14.1
importlib_metadata==6.8.0
importlib_metadata==4.6.4
intel_extension_for_pytorch==2.0.100
ipadic==1.0.0
ipython==8.14.0
jax==0.4.14
jaxlib==0.4.14
jieba==0.42.1
Js2Py==0.74
kenlm==0.1
keras==2.11.0
keras_nlp==0.6.1
matplotlib==3.7.2
matplotlib==3.6.2
megatron==0.5.1
mlflow==2.5.0
more_itertools==8.10.0
msgpack_python==0.5.6
MultiScaleDeformableAttention==1.0
mxnet==1.9.1
natten==0.14.6
neptune==1.4.1
numpy==1.22.0
numpy==1.25.1
onnxruntime==1.15.1
optuna==3.2.0
packaging==21.3
paddleocr==2.6.1.3
paddleocr.egg==info
pandas==2.0.3
pandas==1.5.3
pandas==1.3.5
phonemizer==3.2.1
Pillow==10.0.0
Pillow==9.3.0
Pillow==10.0.0
prettytable==3.8.0
protobuf==3.20.0
psutil==5.9.5
py3nvml==0.2.7
pycocotools==2.0.6
pyctcdecode==0.5.0
pydantic==1.10.7
pyknp==0.6.1
pytesseract==0.3.10
pytest==7.4.0
pythainlp==4.0.2
pytorch_crf==0.7.2
pytorch_lightning==2.0.6
pytorch_quantization==0.0.1.dev5
PyYAML==6.0.1
PyYAML==6.0
PyYAML==5.4.1
PyYAML==6.0.1
ray==2.6.1
regex==2022.10.31
Requests==2.31.0
rjieba==0.1.11
s3prl==0.4.10
sacremoses==0.0.53
safetensors==0.3.1
scann==1.2.10
scikit_learn==1.1.3
scipy==1.11.1
sentence_transformers==2.2.2
sentencepiece==0.1.97
sigopt==8.8.2
spacy==3.6.0
starlette==0.31.0
sudachipy==0.6.7
t5x==0.0.0
tensorboardX==2.6.2
tensorflow==2.13.0
tensorflow_hub==0.14.0
tensorflow_probability==0.20.1
tensorflow_text==2.13.0
tensorstore==0.1.40
tf2onnx==1.14.0
timeout_decorator==0.5.0
timm==0.9.2
tokenizers==0.13.2
torch==1.12.1+cu116
torch_xla==1.0
torchaudio==0.12.1+cu116
torchmetrics==0.11.0
torchvision==0.13.1+cu116
tqdm==4.65.0
tqdm==4.64.1
trajectory==0.1.1
transformers==4.28.1
typing_extensions==4.7.1
unidic==1.1.0
unidic_lite==1.0.8
uvicorn==0.23.2
vissl==0.1.6
wandb==0.15.2
wget==3.2
